produced seemingly correct fused ugraph
