+++ utensor_cgen/transformer/base.py
@@ -62,6 +62,7 @@ class Transformer(object):
       #           for t_info in op_info.input_tensors]
 
       #TODO: move the code below to a standalone function. Consider using a more extensive data structure:
+      #      Or, use this: in_ops = [node.name for node in ugraph.ops_info[op_name].input_nodes]
       tensors_in = set([t.name for t in ugraph.ops_info[op_name].input_tensors])
       in_ops = set()
       for it_node in ugraph.topo_order:
